requirements.txt
asgiref==3.7.2
cloudinary==1.35.0
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==3.2.25
django-allauth==0.54.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
oauthlib==3.3.1
Pillow==10.0.1
psycopg2==2.9.7
psycopg2-binary==2.9.9
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==2.0.0
sqlparse==0.4.4
typing_extensions==4.14.1
urllib3==1.26.20
whitenoise==6.6.0
